Move TransactionId to presto-common
