+++ presto-common/src/main/java/com/facebook/presto/common/transaction/TransactionId.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.transaction;
+package com.facebook.presto.common.transaction;
 
 import com.facebook.drift.annotations.ThriftConstructor;
 import com.facebook.drift.annotations.ThriftField;
