+++ presto-main/src/main/java/com/facebook/presto/connector/system/GlobalSystemTransactionHandle.java
@@ -13,8 +13,8 @@
  */
 package com.facebook.presto.connector.system;
 
+import com.facebook.presto.common.transaction.TransactionId;
 import com.facebook.presto.spi.connector.ConnectorTransactionHandle;
-import com.facebook.presto.transaction.TransactionId;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
