+++ presto-main/src/main/java/com/facebook/presto/connector/system/SystemTransactionHandle.java
@@ -13,9 +13,9 @@
  */
 package com.facebook.presto.connector.system;
 
+import com.facebook.presto.common.transaction.TransactionId;
 import com.facebook.presto.spi.ConnectorId;
 import com.facebook.presto.spi.connector.ConnectorTransactionHandle;
-import com.facebook.presto.transaction.TransactionId;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.Supplier;
