+++ presto-main/src/main/java/com/facebook/presto/server/SessionContext.java
@@ -13,12 +13,12 @@
  */
 package com.facebook.presto.server;
 
+import com.facebook.presto.common.transaction.TransactionId;
 import com.facebook.presto.spi.function.SqlFunctionId;
 import com.facebook.presto.spi.function.SqlInvokedFunction;
 import com.facebook.presto.spi.security.Identity;
 import com.facebook.presto.spi.session.ResourceEstimates;
 import com.facebook.presto.spi.tracing.Tracer;
-import com.facebook.presto.transaction.TransactionId;
 import com.google.common.collect.ImmutableList;
 
 import javax.annotation.Nullable;
