+++ presto-main/src/main/java/com/facebook/presto/transaction/InMemoryTransactionManager.java
@@ -16,6 +16,7 @@ package com.facebook.presto.transaction;
 import com.facebook.airlift.concurrent.BoundedExecutor;
 import com.facebook.airlift.concurrent.ExecutorServiceAdapter;
 import com.facebook.airlift.log.Logger;
+import com.facebook.presto.common.transaction.TransactionId;
 import com.facebook.presto.metadata.Catalog;
 import com.facebook.presto.metadata.CatalogManager;
 import com.facebook.presto.metadata.CatalogMetadata;
