+++ presto-main/src/main/java/com/facebook/presto/transaction/TransactionBuilder.java
@@ -14,6 +14,7 @@
 package com.facebook.presto.transaction;
 
 import com.facebook.presto.Session;
+import com.facebook.presto.common.transaction.TransactionId;
 import com.facebook.presto.security.AccessControl;
 import com.facebook.presto.spi.transaction.IsolationLevel;
