+++ presto-tests/src/test/java/com/facebook/presto/execution/TestingSessionContext.java
@@ -14,6 +14,7 @@
 package com.facebook.presto.execution;
 
 import com.facebook.presto.Session;
+import com.facebook.presto.common.transaction.TransactionId;
 import com.facebook.presto.server.SessionContext;
 import com.facebook.presto.spi.ConnectorId;
 import com.facebook.presto.spi.function.SqlFunctionId;
@@ -21,7 +22,6 @@ import com.facebook.presto.spi.function.SqlInvokedFunction;
 import com.facebook.presto.spi.security.Identity;
 import com.facebook.presto.spi.session.ResourceEstimates;
 import com.facebook.presto.spi.tracing.Tracer;
-import com.facebook.presto.transaction.TransactionId;
 import com.google.common.collect.ImmutableMap;
 
 import java.util.Map;
